Moved MetadataUtil out of util.images into util because functionality is not restricted to images.
